[hotfix][mesos] Remove unused methods and extract interface for MesosArtifactServer.
This makes testing easier for mesos deployment, by allowing using of testing implementation rather than mocking for MesosArtifactServer.
